addressing PR comments, fixing facet filter bug w/ Azul results
