+++ app/javascript/components/search/controls/KeywordSearch.js
@@ -1,10 +1,12 @@
-import React, { useContext, useRef } from 'react'
+import React, { useContext, useRef, useState } from 'react'
 import Button from 'react-bootstrap/lib/Button'
 import InputGroup from 'react-bootstrap/lib/InputGroup'
 import Form from 'react-bootstrap/lib/Form'
-import { faSearch, faTimes } from '@fortawesome/free-solid-svg-icons'
+import { faQuestionCircle, faSearch, faTimes } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { SearchSelectionContext } from 'providers/SearchSelectionProvider'
+import Modal from 'react-bootstrap/lib/Modal'
+import { closeModal } from 'components/search/controls/SearchPanel'
 
 /**
  * Component to search using a keyword value
@@ -17,6 +19,21 @@ export default function KeywordSearch({ keywordPrompt }) {
   //  as long as the text hasn't been updated
   const showClear = selectionContext.terms && selectionContext.terms.length
   const inputField = useRef()
+  const [showTextSearchHelpModal, setTextShowSearchHelpModal] = useState(false)
+
+  const textSearchModalContent = (<div>
+    <h4 className="text-center">Title and description Search</h4><br/>
+    Using the search box below will perform a text-based search of all study titles and descriptions, and will return
+    those that contain any of the specified terms. You can quote (&quot;) phrases to find matches that contain the entire
+    phrase, and you can combine both single terms and quoted phrases in your search.
+  </div>)
+
+  const textSearchLink = <a className="action advanced-opts"
+                            onClick={() => setTextShowSearchHelpModal(true)}
+                            data-analytics-name="search-help">
+    <FontAwesomeIcon icon={faQuestionCircle} />
+  </a>
+
   /**
    * Updates terms in search context upon submitting keyword search
    */
@@ -42,7 +59,9 @@ export default function KeywordSearch({ keywordPrompt }) {
       onSubmit = {handleSubmit}
       className='study-keyword-search'
     >
-      <span className='badge text-search search-title'>Text search</span>
+      <span className='badge text-search search-title'>
+        Description search {textSearchLink}
+      </span>
       <InputGroup>
         <input
           ref = {inputField}
@@ -65,6 +84,15 @@ export default function KeywordSearch({ keywordPrompt }) {
             <FontAwesomeIcon icon={faTimes} />
           </Button> }
       </InputGroup>
+      <Modal
+        show={showTextSearchHelpModal}
+        onHide={() => closeModal(setTextShowSearchHelpModal)}
+        animation={false}
+        bsSize='large'>
+        <Modal.Body className="">
+          { textSearchModalContent }
+        </Modal.Body>
+      </Modal>
     </Form>
   )
 }
