+++ app/javascript/styles/_search.scss
@@ -320,10 +320,8 @@ nav.search-links, #search-panel {
 }
 
 .advanced-opts {
-  vertical-align: top;
-  display: inline-block;
-  padding: 0.65em 0.25em;
-  border-radius: 1em;
+  color: #fff !important;
+  display: inline;
 }
 
 .new-feature {
